Module jemalloc_ctl::opt

Information about the run-time jemalloc configuration.

These settings are controlled by the MALLOC_CONF environment variable.

Structs

Abort
A type determining if jemalloc will call abort(3) on most warnings.

Dss
A type providing access to the dss (sbrk(2)) allocation precedence as related to mmap(2) allocation.

Junk
A type providing access to jemalloc's junk filling mode.

LgTcacheMax
A type providing access to the maximum size class (log base 2) to cache in the thread-specific cache (tcache).

NArenas
A type providing access to the maximum number of arenas to use for automatic multiplexing of threads and arenas.

Tcache
A type providing access to thread-local allocation caching behavior.

Zero
A type providing access to jemalloc's zeroing behavior.

Functions

abort
Determines if jemalloc will call abort(3) on most warnings.

dss
Returns the dss (sbrk(2)) allocation precedence as related to mmap(2) allocation.

junk
Returns jemalloc's junk filling mode.

lg_tcache_max
Returns the maximum size class (log base 2) to cache in the thread-specific cache (tcache).

narenas
Returns the maximum number of arenas to use for automatic multiplexing of threads and arenas.

tcache
Determines if thread-local allocation caching is enabled.

zero
Returns jemalloc's zeroing behavior.