Function jemalloc_ctl::stats::resident

pub fn resident() -> Result<usize>

Returns the total number of bytes in physically resident data pages mapped by the allocator.

This consists of all pages dedicated to allocator metadata, pages backing active allocations, and unused dirty pages. It may overestimate the true value because pages may not actually be physically resident if they correspond to demand-zeroed virtual memory that has not yet been touched. This is a multiple of the page size, and is larger than the value returned by active.

This statistic is cached, and is only refreshed when the epoch is advanced. See the epoch function for more information.

This corresponds to stats.resident in jemalloc's API.

Examples

extern crate jemallocator;
extern crate jemalloc_ctl;

#[global_allocator]
static ALLOC: jemallocator::Jemalloc = jemallocator::Jemalloc;

fn main() {
    jemalloc_ctl::epoch().unwrap();
    println!("{} bytes of total resident data", jemalloc_ctl::stats::resident().unwrap());
}